+++ components/src/core/user-menu/user-menu.component.ts
@@ -4,6 +4,9 @@ import {
     Component,
     EventEmitter,
     Input,
+    OnChanges,
+    OnDestroy,
+    OnInit,
     Output,
     SimpleChanges,
     TemplateRef,
@@ -15,6 +18,7 @@ import { animate, state, style, transition, trigger } from '@angular/animations'
 import { MatBottomSheet } from '@angular/material/bottom-sheet';
 import { fromEvent, Subscription } from 'rxjs';
 import { map, startWith } from 'rxjs/operators';
+import {requireContent, requireInput} from "../../utils/utils";
 
 /**
  * [UserMenu Component](https://pxblue-components.github.io/angular/?path=/info/components-user-menu--readme)
@@ -42,28 +46,28 @@ import { map, startWith } from 'rxjs/operators';
     template: `
         <!-- This button triggers the overlay and is it's origin -->
         <pxb-user-menu-avatar
-                cdkOverlayOrigin
-                [avatarValue]="avatarValue"
-                [avatarImage]="avatarImage"
-                (click)="openOverlay()"
-                #trigger="cdkOverlayOrigin"
+            cdkOverlayOrigin
+            [avatarValue]="avatarValue"
+            [avatarImage]="avatarImage"
+            (click)="openOverlay()"
+            #trigger="cdkOverlayOrigin"
         >
             <ng-content select="[pxb-avatar]"></ng-content>
         </pxb-user-menu-avatar>
 
         <ng-template #menu>
             <pxb-drawer-header
-                    *ngIf="menuTitle"
-                    class="pxb-user-menu-header"
-                    [title]="menuTitle"
-                    [subtitle]="menuSubtitle"
-                    [divider]="true"
+                *ngIf="menuTitle"
+                class="pxb-user-menu-header"
+                [title]="menuTitle"
+                [subtitle]="menuSubtitle"
+                [divider]="true"
             >
                 <pxb-user-menu-avatar
-                        pxb-icon
-                        [avatarValue]="avatarValue"
-                        [avatarImage]="avatarImage"
-                        class="pxb-user-menu-header-avatar"
+                    pxb-icon
+                    [avatarValue]="avatarValue"
+                    [avatarImage]="avatarImage"
+                    class="pxb-user-menu-header-avatar"
                 >
                     <ng-content select="[pxb-menu-avatar]"></ng-content>
                 </pxb-user-menu-avatar>
@@ -73,15 +77,15 @@ import { map, startWith } from 'rxjs/operators';
         </ng-template>
 
         <ng-template
-                cdkConnectedOverlay
-                (backdropClick)="onClickMenuBackdrop()"
-                [cdkConnectedOverlayPush]="true"
-                [cdkConnectedOverlayHasBackdrop]="true"
-                [cdkConnectedOverlayOrigin]="trigger"
-                [cdkConnectedOverlayOpen]="isMenuOpen"
-                [cdkConnectedOverlayPositions]="positions"
-                [cdkConnectedOverlayViewportMargin]="16"
-                [cdkConnectedOverlayBackdropClass]="'pxb-user-menu-overlay-backdrop'"
+            cdkConnectedOverlay
+            (backdropClick)="onClickMenuBackdrop()"
+            [cdkConnectedOverlayPush]="true"
+            [cdkConnectedOverlayHasBackdrop]="true"
+            [cdkConnectedOverlayOrigin]="trigger"
+            [cdkConnectedOverlayOpen]="isMenuOpen"
+            [cdkConnectedOverlayPositions]="positions"
+            [cdkConnectedOverlayViewportMargin]="16"
+            [cdkConnectedOverlayBackdropClass]="'pxb-user-menu-overlay-backdrop'"
         >
             <mat-card class="pxb-user-menu-overlay mat-elevation-z8" [@fade-in-out]>
                 <ng-template [ngTemplateOutlet]="menu"></ng-template>
@@ -92,7 +96,7 @@ import { map, startWith } from 'rxjs/operators';
         class: 'pxb-user-menu',
     },
 })
-export class UserMenuComponent {
+export class UserMenuComponent implements OnInit, OnChanges, OnDestroy {
     /** Image source for avatar */
     @Input() avatarImage: string;
 
@@ -110,11 +114,8 @@ export class UserMenuComponent {
         new ConnectionPositionPair({ originX: 'start', originY: 'top' }, { overlayX: 'start', overlayY: 'top' }),
     ];
 
-    /** Whether the menu overlay appears on screen.
-     *
-     * @default false
-     */
-    @Input() open = false;
+    /** Whether the menu overlay appears on screen. */
+    @Input() open;
 
     /** Window pixel width at which the responsive bottom sheet menu is triggered (set to 0 to disable responsive behavior)
      *
@@ -139,12 +140,13 @@ export class UserMenuComponent {
     checkScreenSize = (): boolean => document.body.offsetWidth < this.useBottomSheetAt;
 
     ngOnInit(): void {
+        requireInput<UserMenuComponent>(['open'], this);
+
         // Subscribe to resize events and transition from menu to bottomsheet (or vise versa) when open & resizing
         this.screenSizeChangeListener = fromEvent(window, 'resize')
             .pipe(map(this.checkScreenSize))
             .pipe(startWith(this.checkScreenSize()))
             .subscribe((isMobile: boolean) => {
-                console.log('hi');
                 // Transition from Desktop to Mobile
                 if (this.open && isMobile && !this.useBottomSheet) {
                     this.isMenuOpen = false;
@@ -159,16 +161,26 @@ export class UserMenuComponent {
             });
     }
 
-    ngOnChanges(simpleChanges: SimpleChanges): void {
+    ngOnChanges(changes: SimpleChanges): void {
         // Set state and dismiss bottom sheet when open() changes.
-        console.log(simpleChanges);
-        if (this.open) {
-            this.openOverlay();
-        } else {
-            this.isMenuOpen = false;
-            this._bottomSheet.dismiss(false);
+        if (changes.open) {
+            // State changes from closed to open.
+            const openState = changes.open;
+            if (openState.currentValue === true && (openState.previousValue === false || openState.isFirstChange())) {
+                if (this.useBottomSheet) {
+                    this._openBottomSheet();
+                } else {
+                    this.isMenuOpen = true;
+                }
+            }
+            // State changes from open to closed.
+            if (openState.currentValue === false && (openState.previousValue === true || openState.isFirstChange())) {
+                this.isMenuOpen = false;
+                this._bottomSheet.dismiss(false);
+            }
         }
-        if (simpleChanges.useBottomSheetAt) {
+
+        if (changes.useBottomSheetAt) {
             this.useBottomSheet = this.checkScreenSize();
         }
     }
@@ -180,40 +192,33 @@ export class UserMenuComponent {
     }
 
     onClickMenuBackdrop(): void {
-        this.open = false;
-        this.isMenuOpen = false;
-        this.openChange.emit(this.open);
+        this.openChange.emit(false);
         this.backdropClick.emit();
     }
 
     openOverlay(): void {
-        console.log('open overlay');
-        this.open = true;
-        this.openChange.emit(this.open);
-        if (this.useBottomSheet) {
-            this._openBottomSheet();
-        } else {
-            this.isMenuOpen = true;
-        }
+        this.openChange.emit(true);
     }
 
     private _openBottomSheet(): void {
-        const ref = this._bottomSheet.open(this.menu, {
+
+        // Do not invoke the bottom sheet until `menu` has been rendered.
+        this._ref.detectChanges();
+
+        const bottomSheetRef = this._bottomSheet.open(this.menu, {
             backdropClass: 'pxb-user-menu-bottomsheet-backdrop',
             panelClass: 'pxb-user-menu-bottomsheet',
             hasBackdrop: true,
         });
-        ref.afterDismissed()
-            .subscribe((openMenu: true) => {
-                if (openMenu) {
-                    this.isMenuOpen = true;
-                    this._ref.detectChanges();
-                }
-            });
 
-        ref.backdropClick().subscribe(() => {
-            this.backdropClick.emit();
+        bottomSheetRef.afterDismissed().subscribe((openMenu: true) => {
+            this.isMenuOpen = openMenu;
+            this._ref.detectChanges();
         });
 
+        bottomSheetRef.backdropClick().subscribe(() => {
+            this.openChange.emit(false);
+            this.backdropClick.emit();
+        });
     }
 }