+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Change Log
 
 ## v6.0.0 (Unreleased)
+
 ### Added
+
 -   Added new property `unitSpace` to `<pxb-channel-value>` to manage spacing between the value and units.
+
+### Fixed
+
+-   Fixed stepper spacing in `<pxb-mobile-stepper` when Back and Next buttons are uneven width.  
+
 ## v5.0.1 (July 30, 2021)
 
 ### Added