+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 -   Fixed stepper spacing in `<pxb-mobile-stepper>` when Back and Next buttons are uneven width.  
 -   Fixed bug in `<pxb-app-bar>` that prevented class overrides on the root element.
+-   Fixed bug in `<pxb-app-bar>` so scroll listeners can no longer attempt to measure `undefined` elements' scroll distance. 
 
 ## v5.0.1 (July 30, 2021)